@@ -1,54 +1,40 @@
 {
 	"name": "portfolio",
-	"version": "0.1.0",
 	"private": true,
-	"dependencies": {
-		"@apollo/client": "^3.3.15",
-		"@craco/craco": "^6.1.1",
-		"@fortawesome/fontawesome-svg-core": "^1.2.35",
-		"@fortawesome/free-brands-svg-icons": "^5.15.3",
-		"@fortawesome/free-solid-svg-icons": "^5.15.3",
-		"@fortawesome/react-fontawesome": "^0.1.14",
-		"@testing-library/jest-dom": "^5.11.4",
-		"@testing-library/react": "^11.1.0",
-		"@testing-library/user-event": "^12.1.10",
-		"graphql": "^15.5.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-markdown": "^6.0.1",
-		"react-photo-gallery": "^8.0.0",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "4.0.3",
-		"web-vitals": "^1.0.1"
-	},
+	"version": "1.0.0",
+	"type": "module",
 	"scripts": {
-		"start": "craco start",
-		"build": "craco build",
-		"test": "craco test",
-		"eject": "react-scripts eject"
+		"dev": "vite",
+		"build": "tsc && vite build",
+		"lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+		"preview": "vite preview"
 	},
-	"eslintConfig": {
-		"extends": [
-			"react-app",
-			"react-app/jest"
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
+	"dependencies": {
+		"@apollo/client": "^3.8.10",
+		"@fortawesome/fontawesome-svg-core": "^6.5.1",
+		"@fortawesome/free-brands-svg-icons": "^6.5.1",
+		"@fortawesome/free-solid-svg-icons": "^6.5.1",
+		"@fortawesome/react-fontawesome": "^0.2.0",
+		"graphql": "^16.8.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-markdown": "^9.0.1",
+		"react-photo-gallery": "^8.0.0",
+		"react-router-dom": "^6.21.3"
 	},
 	"devDependencies": {
-		"@tailwindcss/postcss7-compat": "^2.1.0",
-		"autoprefixer": "^9.8.6",
-		"postcss": "^7.0.35",
-		"tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.0"
+		"@types/react": "^18.2.43",
+		"@types/react-dom": "^18.2.17",
+		"@typescript-eslint/eslint-plugin": "^6.14.0",
+		"@typescript-eslint/parser": "^6.14.0",
+		"@vitejs/plugin-react": "^4.2.1",
+		"autoprefixer": "^10.4.17",
+		"eslint": "^8.55.0",
+		"eslint-plugin-react-hooks": "^4.6.0",
+		"eslint-plugin-react-refresh": "^0.4.5",
+		"postcss": "^8.4.33",
+		"tailwindcss": "^3.4.1",
+		"typescript": "^5.2.2",
+		"vite": "^5.0.8"
 	}
 }
