@@ -1,24 +1,24 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 import { ApolloProvider } from "@apollo/client";
 import WithTheme from "./WithTheme";
 import { CLIENT } from "./graphql/apollo";
 
 const AppWithTheme = WithTheme(App);
 
-ReactDOM.render(
-	<React.StrictMode>
+const root = createRoot(document.getElementById("root")!);
+
+root.render(
+	<StrictMode>
 		<ApolloProvider client={CLIENT}>
 			<AppWithTheme />
 		</ApolloProvider>
-	</React.StrictMode>,
-	document.getElementById("root")
+	</StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+// reportWebVitals();
