@@ -1,10 +1,10 @@
-import React, { useState, useContext } from "react";
+import { useState, useContext } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSun, faMoon } from "@fortawesome/free-solid-svg-icons";
-import ThemeContext from "./../../ThemeContext";
+import ThemeContext from "../../ThemeContext";
 
-import { useContentful } from "./../../Hooks/useContentful";
-import { THEME_QUERY } from "./../../graphql/queries";
+import { useContentful } from "../../Hooks/useContentful";
+import { THEME_QUERY } from "../../graphql/queries";
 
 function DarkModeButton() {
 	const { gettheme, toggleTheme } = useContext(ThemeContext);
