@@ -1,9 +1,14 @@
-import React from "react";
 import "./Lightbox.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimes } from "@fortawesome/free-solid-svg-icons";
 
-function Lightbox({ image, open, close }) {
+interface LightboxInterface {
+	image: string;
+	open: boolean;
+	close: any;
+}
+
+function Lightbox({ image, open, close }: LightboxInterface) {
 	return (
 		<dialog className="lightbox" open={open} onClick={close}>
 			<div className="background"></div>
