@@ -1,9 +1,14 @@
-import React from "react";
 import "./LoadingScreen.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSpinner } from "@fortawesome/free-solid-svg-icons";
 
-function LoadingScreen({ screen, minHeight, times }) {
+interface LoadinScreenInterface {
+	screen?: boolean;
+	minHeight?: string;
+	times?: number;
+}
+
+function LoadingScreen({ screen, minHeight, times }: LoadinScreenInterface) {
 	return (
 		<>
 			{screen ? (
@@ -11,6 +16,7 @@ function LoadingScreen({ screen, minHeight, times }) {
 					<FontAwesomeIcon icon={faSpinner} spin />
 				</section>
 			) : (
+				//@ts-ignore
 				[...Array(times)].map((i, index) => {
 					return (
 						<section
