@@ -1,11 +1,24 @@
-import React, { useEffect } from "react";
+import { useEffect } from "react";
 import "./Education.css";
 
-import LoadingScreen from "./../../components/loadingScreen/LoadingScreen";
+import LoadingScreen from "../../components/loadingScreen/LoadingScreen";
 import EducationCard from "./EducationCard";
 
-import { useContentful } from "./../../Hooks/useContentful";
-import { EDUCATION_QUERY } from "./../../graphql/queries";
+import { useContentful } from "../../Hooks/useContentful";
+import { EDUCATION_QUERY } from "../../graphql/queries";
+
+interface EducationInterface {
+	sys: {
+		id: string;
+	}
+	title: string;
+	date: string;
+	subtitle: string;
+	image?: {
+		url: string;
+		alt?: string;
+	}
+}
 
 function Education() {
 	const [content, loading] = useContentful(EDUCATION_QUERY);
@@ -27,7 +40,7 @@ function Education() {
 						<header>
 							<h2>{content.education.heading}</h2>
 						</header>
-						{content?.degreeCollection?.items.map((degree) => {
+						{content?.degreeCollection?.items.map((degree: EducationInterface) => {
 							return (
 								<EducationCard
 									key={degree.sys.id}
@@ -40,7 +53,7 @@ function Education() {
 						<header>
 							<h2>{content.certification.heading}</h2>
 						</header>
-						{content?.educationCollection?.items.map((cert) => {
+						{content?.educationCollection?.items.map((cert: EducationInterface) => {
 							return (
 								<EducationCard key={cert.sys.id} info={cert} />
 							);
