@@ -1,13 +1,24 @@
-import React from "react";
+interface EducationCardInterface {
+	info: {
+		title: string;
+		date: string;
+		subtitle: string;
+		image?: {
+			url: string;
+			alt?: string;
+		}
+	}
+}
 
-function EducationCard({ info }) {
+function EducationCard({ info }: EducationCardInterface) {
 	return (
 		<section className="education_section">
-			<img
+			{ info.image? <img
 				src={info.image.url + "?fl=progressive&w=200"}
 				alt=""
 				loading="lazy"
-			/>
+			/> : ''}
+			
 			<section>
 				<h3>
 					{info.title}&nbsp;
